+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.38.0] - 2025-06-24
+
 ### Added
 
 - `tilebox-datasets`: Added `delete_collection` method to `DatasetClient` to delete a collection.
+- `tilebox-workflows`: Added support for nested dataclasses and protobuf messages inside a `list`, `dict` or `tuple`
+  as task arguments.
 
 ## [0.37.1] - 2025-06-10
 
@@ -194,7 +198,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.37.1...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.38.0...HEAD
+[0.38.0]: https://github.com/tilebox/tilebox-python/compare/v0.37.1...v0.38.0
 [0.37.1]: https://github.com/tilebox/tilebox-python/compare/v0.37.0...v0.37.1
 [0.37.0]: https://github.com/tilebox/tilebox-python/compare/v0.36.1...v0.37.0
 [0.36.1]: https://github.com/tilebox/tilebox-python/compare/v0.36.0...v0.36.1